+++ roguelike_crafter/Assets/New Terrain 1.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 119a09ebceacdec47920c08321982cf7
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 15600000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
