+++ roguelike_crafter/Assets/New Terrain 2.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8ab53f4baf4886e46ad22a1acc8c38a2
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 15600000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
