--- roguelike_crafter/Assets/Scripts/FallOffGenerator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: df79fa85d4c41634c8b88f65bf4a2f9f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
