+++ roguelike_crafter/Assets/Scripts/FloorBox.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 49e82049878ff094ba12ecda4216bfe4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
