+++ roguelike_crafter/Assets/Scripts/LookingCam.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 51e5bde57e4f7b947a998b6b9895c282
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
