--- roguelike_crafter/Assets/Scripts/MapCreaEditor.cs
@@ -1,24 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEditor;
-[CustomEditor (typeof (MapCreation))]
-public class MapCreaEditor : Editor
-{
-    public override void OnInspectorGUI()
-    {
-        MapCreation mapGen = (MapCreation) target;
-        if (DrawDefaultInspector())
-        {
-            if (mapGen.autoUpdate)
-            {
-                mapGen.GenerateMap();
-            }   
-        }
-        if (GUILayout.Button("Generate"))
-        {
-            mapGen.GenerateMap();
-        }
-    }
-    
-}
