--- roguelike_crafter/Assets/Scripts/MapCreaEditor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8ef5b5d66b173f84d80a988302e6ad0c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
