--- roguelike_crafter/Assets/Scripts/MapCreation.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f61dc84958164fc4682416ca205d7d85
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
