--- roguelike_crafter/Assets/Scripts/MapDisplay.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3d2f8d06a50435f4da496ce76599bc6c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
