+++ roguelike_crafter/Assets/Scripts/Scrapped Idea.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b7d0ec8eb1a6915478674b944d5da39e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
