+++ roguelike_crafter/Assets/Scripts/Scrapped Idea/FallOffGenerator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: df79fa85d4c41634c8b88f65bf4a2f9f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
