+++ roguelike_crafter/Assets/Scripts/Scrapped Idea/FallOffGenerator1.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c7b1003f5d761874181e78e0f994248a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
