+++ roguelike_crafter/Assets/Scripts/Scrapped Idea/MapCreation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f61dc84958164fc4682416ca205d7d85
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
