+++ roguelike_crafter/Assets/Scripts/Scrapped Idea/MeshGen.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6b1a4864b63a952409cfe2ad309bb278
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
