+++ roguelike_crafter/Assets/Scripts/Scrapped Idea/Noise.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5d5608139a10c9b46a9c1a567959a7f7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
