+++ roguelike_crafter/Assets/Scripts/Scrapped Idea/TextGen.cs
@@ -1,0 +1,36 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public static class TextGen
+{
+    public static Texture2D textureFromCM(Color[] colorMap, int w, int h)
+    {
+        Texture2D texture = new Texture2D(w, h);
+        texture.filterMode = FilterMode.Point;
+        texture.wrapMode = TextureWrapMode.Clamp;
+        texture.SetPixels(colorMap);
+        texture.Apply();
+        return texture;
+    }
+
+    public static Texture2D textureFromHM(float[,] heightMap)
+    {
+     int width = heightMap.GetLength(0);
+     int height = heightMap.GetLength(1);
+
+       Texture2D texture2D = new Texture2D(width, height);
+       Color[] colorMap = new Color[width * height];
+
+       for (int y = 0; y < height; y++)
+       {
+           for (int x = 0; x < width; x++)
+           {
+               colorMap[y * width + x] = Color.Lerp(Color.black, Color.white, heightMap[x, y]);
+
+           }
+       }
+       return textureFromCM(colorMap, width, height);
+    }
+}
+
