+++ roguelike_crafter/Assets/Scripts/Scrapped Idea/TextGen.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2a515aa815e2178429bdbdc4396d421f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
