--- roguelike_crafter/Assets/Scripts/TextGen.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2a515aa815e2178429bdbdc4396d421f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
