+++ roguelike_crafter/Assets/TerrainData_1d930ab4-e108-4715-ba3d-fcc53fd483b2.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9fe7312918b0c8e44a3c0e88c020e9f2
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 15600000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
