+++ roguelike_crafter/Assets/TerrainData_96cf28fc-18c2-48e4-a3d6-47f9fd0e51ce.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a3701e1c65146294b980cbdc9bcf7912
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 15600000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
