+++ roguelike_crafter/ProjectSettings/TagManager.asset
@@ -8,7 +8,7 @@ TagManager:
   - Default
   - TransparentFX
   - Ignore Raycast
-  - 
+  - Ground
   - Water
   - UI
   -
